chore: switch to Elementary Pro
