+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "homepage": "https://teetow.github.io/elementary_grid",
   "dependencies": {
-    "@nick-thompson/elementary": "^0.10.4",
+    "@elemaudio/core": "^0.11.0",
     "@types/classnames": "^2.3.1",
     "@types/nick-thompson__elementary": "^0.9.0",
     "@types/node": "^16.11.7",
