--- src/lib/elementary.jsx
@@ -1,7 +1,0 @@
-import {
-  el,
-  ElementaryWebAudioRenderer as core,
-} from "@nick-thompson/elementary";
-
-export { el, core };
-export default core;
